Rounding p2

+++ streamlit_app.py
@@ -37,8 +37,8 @@ if uploaded_file is not None:
     # Identify peaks in the smoothed data
     peaks, _ = find_peaks(smooth_data, height=0.05, width=10)
 
-    # Round the peak absorbances to 4 decimal places
-    rounded_absorbances = [round(absorbance, 4) for absorbance in smooth_data[peaks]]
+    # Round the peak absorbances to 4 decimal places and convert to regular Python floats
+    rounded_absorbances = [float(round(absorbance, 4)) for absorbance in smooth_data[peaks]]
 
     # Debugging: Output the peak indices and values with reduced decimals
     st.write(f"Peaks found at indices: {peaks.tolist()}")
